fix(ci): 修复 gofmt/lint 积压与门禁失守——全量检查链回绿
- gofmt：pool_http_test.go / clientinvoke.go 格式漂移；此前 CI 卡死在
  gofmt，lint/test/codegen 等后续门禁自 09-11 起从未执行，问题层层积压
- golangci-lint 25 条清零：rows.Close errcheck 统一 defer func(){_ = ...}；
  G115 按分页/字典软上限口径加行内豁免并注明理由；G301/G302/G304 豁免
  注明用户源文件/测试夹具语义；unused/ineffassign 清理（pruneBatchLimit、
  observeEgressClass、currentPeriod 等 3 处死码）
- setTestUmask 改返回恢复函数：修正「set 后立即 restore」等于未生效
  （umask 0077 回归形态此前从未真正复现）；两处调用点同步
- console 时区测试：匿名场景期望对齐包内约定 PermissionDenied
  （该测试自提交起从未通过）
- config.pb.go 用钉版 protoc v6.31.1 重新生成（提交时为 4.23.4，
  codegen drift 门禁必红）
- 验证：go vet / golangci-lint v2.12.2(Linux 容器) 0 issue / 全量
  go test -race 通过（coverage 56.6% ≥ 48）/ RLS 基准 / buf lint

+++ internal/app/analytics/maintenance_integration_test.go
@@ -24,13 +24,13 @@ type maintenanceEnv struct {
 	query       *appanalytics.Query
 }
 
-func setupMaintenanceEnv(t *testing.T, retentionDays int) *maintenanceEnv {
+func setupMaintenanceEnv(t *testing.T, retentionDays int32) *maintenanceEnv {
 	t.Helper()
 	if testing.Short() {
 		t.Skip("skipping integration test")
 	}
 	e := setupRollupEnv(t)
-	cfg := &config.AppConfig{Analytics: &config.Analytics{RetentionDays: int32(retentionDays)}}
+	cfg := &config.AppConfig{Analytics: &config.Analytics{RetentionDays: retentionDays}}
 	m := appanalytics.NewMaintenanceFromConfig(cfg, e.workerRepo, e.workerRepo, e.projectsRepo, nil)
 	query := appanalytics.NewQueryFromConfig(cfg, bunrepo.NewAnalyticsQueryRepository(e.db))
 	return &maintenanceEnv{rollupEnv: e, maintenance: m, query: query}
